refactor: extract ImageViewerDialog as shared component
- Create common ImageViewerDialog component with download support
- Support single image or multiple images with navigation
- iOS Safari compatible download with blob-based approach
- Configurable title, navigation, and metadata display
- Update JobHistory to use common component
- Update PromptDataList to use common component
