+++ frontend/src/pages/JobHistory.js
@@ -53,6 +53,7 @@ import { jobAPI, workboardAPI, promptDataAPI } from '../services/api';
 import config from '../config';
 import Pagination from '../components/common/Pagination';
 import ImageSelectDialog from '../components/common/ImageSelectDialog';
+import ImageViewerDialog from '../components/common/ImageViewerDialog';
 import { useForm, Controller } from 'react-hook-form';
 
 function SavePromptDialog({ open, onClose, job, onSave }) {
@@ -654,142 +655,6 @@ function JobCard({ job, onView, onRetry, onCancel, onDelete, onImageView, onCont
   );
 }
 
-// 이미지 뷰어 다이얼로그 컴포넌트
-function ImageViewerDialog({ images, selectedIndex, open, onClose }) {
-  const [currentIndex, setCurrentIndex] = useState(selectedIndex || 0);
-
-  React.useEffect(() => {
-    if (selectedIndex !== undefined) {
-      setCurrentIndex(selectedIndex);
-    }
-  }, [selectedIndex]);
-
-  if (!images || images.length === 0) return null;
-
-  const currentImage = images[currentIndex];
-  
-  const handleDownload = async () => {
-    try {
-      // Safari 호환성을 위한 다운로드 방식 개선
-      const response = await fetch(currentImage.url);
-      const blob = await response.blob();
-      
-      // Safari에서 blob URL 생성
-      const blobUrl = window.URL.createObjectURL(blob);
-      
-      // iPhone Safari에서 이미지를 새 탭에서 열어 수동 다운로드 유도
-      const isIOS = /iPad|iPhone|iPod/.test(navigator.userAgent);
-      const isSafari = /Safari/.test(navigator.userAgent) && !/Chrome/.test(navigator.userAgent);
-      
-      if (isIOS && isSafari) {
-        // iOS Safari에서는 새 창으로 이미지를 열어 장기간 누르기로 다운로드 유도
-        const newWindow = window.open(blobUrl, '_blank');
-        if (!newWindow) {
-          // 팝업이 차단된 경우 현재 창에서 열기
-          window.location.href = blobUrl;
-        }
-        toast.success('이미지를 길게 눌러서 저장하세요');
-      } else {
-        // 다른 브라우저에서는 기존 방식 사용
-        const link = document.createElement('a');
-        link.href = blobUrl;
-        link.download = currentImage.originalName || `image_${currentIndex + 1}.png`;
-        document.body.appendChild(link);
-        link.click();
-        document.body.removeChild(link);
-        toast.success('다운로드 완료');
-      }
-      
-      // 메모리 누수 방지
-      setTimeout(() => {
-        window.URL.revokeObjectURL(blobUrl);
-      }, 1000);
-      
-    } catch (error) {
-      console.error('Download error:', error);
-      toast.error('다운로드 실패. 잠시 후 다시 시도해주세요.');
-    }
-  };
-
-  return (
-    <Dialog 
-      open={open} 
-      onClose={onClose} 
-      maxWidth="lg" 
-      fullWidth
-      PaperProps={{
-        sx: { bgcolor: 'black', maxHeight: '90vh' }
-      }}
-    >
-      <DialogTitle sx={{ color: 'white', pb: 1 }}>
-        <Box display="flex" justifyContent="space-between" alignItems="center">
-          <Typography variant="h6">
-            생성된 이미지 ({currentIndex + 1} / {images.length})
-          </Typography>
-          <Box>
-            <IconButton onClick={handleDownload} sx={{ color: 'white', mr: 1 }}>
-              <Download />
-            </IconButton>
-            <IconButton onClick={onClose} sx={{ color: 'white' }}>
-              <Close />
-            </IconButton>
-          </Box>
-        </Box>
-      </DialogTitle>
-      <DialogContent sx={{ textAlign: 'center', p: 2, bgcolor: 'black' }}>
-        <img
-          src={currentImage.url}
-          alt={`Generated ${currentIndex + 1}`}
-          style={{
-            maxWidth: '100%',
-            maxHeight: '70vh',
-            objectFit: 'contain',
-            borderRadius: '8px'
-          }}
-        />
-        
-        {/* 이미지 정보 */}
-        {currentImage.metadata && (
-          <Box mt={2} sx={{ color: 'white' }}>
-            <Typography variant="body2">
-              크기: {currentImage.metadata.width} x {currentImage.metadata.height}
-            </Typography>
-            {currentImage.size && (
-              <Typography variant="body2">
-                파일 크기: {(currentImage.size / 1024 / 1024).toFixed(2)} MB
-              </Typography>
-            )}
-          </Box>
-        )}
-      </DialogContent>
-      
-      {/* 이미지 네비게이션 */}
-      {images.length > 1 && (
-        <DialogActions sx={{ bgcolor: 'black', justifyContent: 'center', pb: 2 }}>
-          <Box display="flex" gap={1} maxWidth="100%" sx={{ overflowX: 'auto' }}>
-            {images.map((image, index) => (
-              <Avatar
-                key={index}
-                src={image.url}
-                onClick={() => setCurrentIndex(index)}
-                sx={{
-                  width: 60,
-                  height: 60,
-                  cursor: 'pointer',
-                  border: index === currentIndex ? '2px solid white' : 'none',
-                  opacity: index === currentIndex ? 1 : 0.7,
-                  '&:hover': { opacity: 1 }
-                }}
-                variant="rounded"
-              />
-            ))}
-          </Box>
-        </DialogActions>
-      )}
-    </Dialog>
-  );
-}
-
 function JobDetailDialog({ job, open, onClose, onImageView }) {
   if (!job) return null;
 
@@ -1281,6 +1146,7 @@ function JobHistory() {
         selectedIndex={selectedImageIndex}
         open={imageViewerOpen}
         onClose={() => setImageViewerOpen(false)}
+        title="생성된 이미지"
       />
 
       <SavePromptDialog
